A PROGRAM SYNTHESIZER BASED ON NUMERICAL INPUT-OUTPUT EXAMPLES
Since the birth of the first electronic computer, computer programs
have helped us accomplished all kinds of hard and difficult tasks,
leaving us only convenience. Nevertheless, the coding work always
puzzles many programmers. We hope that computers could help us generate
programs automatically, setting us free from the boring coding work.
Program synthesis may help us realize that dream. It also does very
good job in the discovery of new algorithm, auto-solving, intelligent
torturing and so on. Moreover basing on the input-output examples is
the most natural and simplest form to present user intents. I made a
simple PROGRAM SYNTHESIZER BASED ON NUMERICAL INPUT-OUTPUT EXAMPLES by
the language of C++. It contains two engines, one of which is a
synthesis engine using the program synthesis technique, which can
discover the regular pattern between the terms; the other is especially
used for calculate the General Term of the polynomial expression. The
former is the focus of my work, and the latter is only for reference
and comparison. When applied to the numerical reasoning problems of the
Civil Service Examinations, it works quite well and is able to solve
most of them. This paper illustrates the details of my realization of
the PROGRAM SYNTHESIZER BASED ON NUMERICAL INPUT-OUTPUT EXAMPLES, the
related techniques, the problems encountered and solutions to them.